pyproject.toml
[project]
name = "ufo-desktop"
version = "0.1.0"
description = "UFO floating desktop character for macOS"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "Pillow>=10.0",
]

[tool.uv]
package = false
